2.1 stable

+++ system/tastyigniter/models/Extensions_model.php
@@ -52,7 +52,7 @@ class Extensions_model extends TI_Model {
 			$extension['settings'] = ! empty($extension_meta['settings'])
 			AND file_exists($extension_path . '/controllers/admin_' . $basename . '.php') ? TRUE : FALSE;
 
-			$extension['config'] = ( ! empty($config) AND is_array($config)) ? TRUE : $config;
+			$extension['config'] = $config;
 
 			$extension['meta'] = ( ! empty($extension_meta) AND is_array($extension_meta)) ? $extension_meta : array();
 
@@ -108,11 +108,11 @@ class Extensions_model extends TI_Model {
 		return $result;
 	}
 
-	public function getExtension($name = '') {
+	public function getExtension($name = '', $filter = array('filter_status' => '1')) {
 		$result = array();
 
 		if ( ! empty($name)) {
-			$extensions = $this->getList(array('filter_status' => '1'));
+			$extensions = $this->getList($filter);
 
 			if ($extensions AND is_array($extensions)) {
 				if (isset($extensions[$name]) AND is_array($extensions[$name])) {
@@ -151,7 +151,7 @@ class Extensions_model extends TI_Model {
 				$row['ext_data'] = ($row['serialized'] === '1' AND ! empty($row['data'])) ? unserialize($row['data']) : array();
 				unset($row['data']);
 				$row['title'] = ! empty($row['title']) ? $row['title'] : ucwords(str_replace('_module', '',
-				                                                                             $row['name']));
+																							 $row['name']));
 				$result[$row['name']] = $row;
 			}
 		}
@@ -165,6 +165,22 @@ class Extensions_model extends TI_Model {
 
 	public function getPayments() {
 		return $this->getInstalledExtensions('payment', TRUE);
+	}
+
+	public function getModule($name = '') {
+		$result = array();
+
+		if ( ! empty($name) AND is_string($name)) {
+			$extensions = $this->getInstalledExtensions('module', TRUE);
+
+			if ($extensions AND is_array($extensions)) {
+				if (isset($extensions[$name]) AND is_array($extensions[$name])) {
+					$result = $extensions[$name];
+				}
+			}
+		}
+
+		return $result;
 	}
 
 	public function getPayment($name = '') {
@@ -278,7 +294,7 @@ class Extensions_model extends TI_Model {
 			if (is_dir($filepath)) {
 				$files = $this->getExtensionFiles($filename, $files);
 			} else {
-				$files[] = $filename;
+				$files[] = EXTPATH . $filename;
 			}
 		}
 
@@ -361,7 +377,7 @@ class Extensions_model extends TI_Model {
 		return $query;
 	}
 
-	public function delete($type = '', $name = '') {
+	public function delete($type = '', $name = '', $delete_data = TRUE) {
 		$query = FALSE;
 
 		if ( ! empty($type) AND $this->extensionExists($name)) {
@@ -375,17 +391,19 @@ class Extensions_model extends TI_Model {
 				$query = TRUE;
 			}
 
-			$this->db->where('status', '0');
-			$this->db->where('type', $type);
-			$this->db->where('name', $name);
+			if ($delete_data) {
+				$this->db->where('status', '0');
+				$this->db->where('type', $type);
+				$this->db->where('name', $name);
 
-			$this->db->delete('extensions');
-			if ($this->db->affected_rows() > 0) {
+				$this->db->delete('extensions');
+				if ($this->db->affected_rows() > 0) {
 
-				// downgrade extension migration
-				$this->extension->runMigration($name, TRUE);
+					// downgrade extension migration
+					$this->extension->runMigration($name, TRUE);
 
-				$query = TRUE;
+					$query = TRUE;
+				}
 			}
 		}
